// lib/api/generated/models/utils-gig-status.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Huerray Backend
 * Comprehensive API for connecting content creators with brands for marketing campaigns and gig opportunities
 *
 * The version of the OpenAPI document: 1.0
 * Contact: support@thrypes.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */




export const UtilsGigStatus = {
    GigStatusDraft: 'draft',
    GigStatusValidated: 'validated',
    GigStatusOpen: 'open',
    GigStatusInProgress: 'in_progress',
    GigStatusReturned: 'returned',
    GigStatusCompleted: 'completed'
} as const;

export type UtilsGigStatus = typeof UtilsGigStatus[keyof typeof UtilsGigStatus];